x86: atomic64_read should return long long
Causes warnings when used as a printk parameter.  All other variants of
the function return long long.

+++ arch/x86/include/asm/atomic64_64.h
@@ -16,7 +16,7 @@
  * Atomically reads the value of @v.
  * Doesn't imply a read memory barrier.
  */
-static inline long atomic64_read(const atomic64_t *v)
+static inline long long atomic64_read(const atomic64_t *v)
 {
 	return (*(volatile long *)&(v)->counter);
 }
